+++ rtp_llm/models_py/triton_kernels/sparse_mla/block_index_to_global.py
@@ -29,7 +29,7 @@ def _convert_req_index_to_global_index_kernel(
 ):
     # program_id(0) -> token_id (row)
     # program_id(1) -> tile index along columns
-    token_id = tl.program_id(0)
+    token_id = tl.program_id(0).to(tl.int64)
     tile_id = tl.program_id(1)
 
     # Each program covers BLOCK_N consecutive columns
